+++ CHANGELOG.md
@@ -13,6 +13,7 @@ are written using raw sql.
 ### Fixed
 - Fixed error drawer from opening on each query change
 - Fixed issue where chart label would be partially cut off
+- Fixed issue where saving a chart would run the query and expand the chart section
 
 ## [0.6.4] - 2021-08-23
 ### Added